refactor: Extract attachIds
